requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0
python-multipart
# Database
psycopg2-binary
sqlalchemy
asyncpg
# Vector DB
qdrant-client
llama-index-core
llama-index-vector-stores-qdrant
llama-index-embeddings-ollama
llama-index-readers-file
